+++ assets/javascripts/kitten/components/form/field/components/checkbox.js
@@ -1,4 +1,8 @@
 import React from 'react'
 import { Checkbox } from '../../../form/checkbox'
 
-export const FieldCheckbox = props => <Checkbox {...props} />
+export const FieldCheckbox = props => (
+  <div className="k-Field__control">
+    <Checkbox {...props} />
+  </div>
+)