+++ assets/javascripts/kitten/components/form/field/components/help.js
@@ -1,20 +1,46 @@
 import React from 'react'
 import { Text } from '../../../typography/text'
+import classNames from 'classnames'
+import { createGlobalStyle } from 'styled-components'
 
-export const FieldHelp = ({ children, ...others }) => {
+const FieldHelpGlobalStyles = createGlobalStyle`
+  .k-Field__control + .k-Field__help--hiddenOnBlur,
+  .k-Field__control + .k-Field__error + .k-Field__help--hiddenOnBlur {
+    opacity: 0;
+    transition: opacity var(--transition);
+  }
+  .k-Field__control:focus-within + .k-Field__help--hiddenOnBlur,
+  .k-Field__control:focus-within + .k-Field__error + .k-Field__help--hiddenOnBlur {
+    opacity: 1;
+  }
+`
+
+export const FieldHelp = ({
+  children,
+  className,
+  hiddenOnBlur = false,
+  ...others
+}) => {
   return (
-    <div className="k-u-margin-top-single">
+    <>
+      <FieldHelpGlobalStyles />
       <Text
         tag="p"
-        color="var(--color-grey-700)"
+        cssColor="var(--color-grey-700)"
         size="micro"
         weight="light"
         lineHeight="normal"
-        className="k-u-margin-none"
+        className={classNames(
+          'k-Field__help',
+          className,
+          'k-u-margin-none',
+          'k-u-margin-top-single',
+          { 'k-Field__help--hiddenOnBlur': hiddenOnBlur },
+        )}
         {...others}
       >
         {children}
       </Text>
-    </div>
+    </>
   )
 }