+++ assets/javascripts/kitten/components/form/field/components/radio-set.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { RadioSet } from '../../../form/radio-set'
 
 export const FieldRadioSet = props => (
-  <div className="k-u-margin-top-single">
+  <div className="k-u-margin-top-single k-Field__control">
     <RadioSet {...props} />
   </div>
 )